feat: Add ES6 Promise module
This commit adds a new ES6 Promise module that exports a function called getResponseFromAPI. This function returns a Promise that resolves after a 2-second delay. If the API call is successful, the Promise resolves with a value of true. Otherwise, it rejects with a value of false.

The new module is added in the ES6_promise directory, with the main file being 0-promise.js. Additionally, a README.md file is added to provide a brief description of the ES6 Promises.