feat: Add handleProfileSignup function
The commit adds a new function called handleProfileSignup to the ES6_promise/3-all.js file. This function uses the ES6 Promise.all method to upload a photo and create a user simultaneously. It then logs the photo body, user's first name, and last name. If there is an error, it logs that the signup system is offline.