feat: Add signUpUser function
The commit adds a new function called signUpUser to the ES6_promise/4-user-promise.js file. This function exports a default function that takes in the user's first name and last name as parameters and returns a Promise that resolves with an object containing the user's first name and last name.